+++ src/main/java/net/p3pp3rf1y/sophisticatedcore/client/ClientEventHandler.java
@@ -21,8 +21,6 @@ import net.minecraft.world.inventory.Slot;
 import net.minecraft.world.inventory.tooltip.TooltipComponent;
 import net.minecraft.world.item.ItemStack;
 import net.p3pp3rf1y.sophisticatedcore.api.IStashStorageItem;
-import net.p3pp3rf1y.sophisticatedcore.client.gui.StorageScreenBase;
-import net.p3pp3rf1y.sophisticatedcore.client.gui.utils.GuiHelper;
 import net.p3pp3rf1y.sophisticatedcore.client.gui.utils.TranslationHelper;
 import net.p3pp3rf1y.sophisticatedcore.client.init.ModFluids;
 import net.p3pp3rf1y.sophisticatedcore.client.init.ModParticles;
@@ -77,11 +75,11 @@ public class ClientEventHandler implements ClientModInitializer {
         AbstractContainerMenu menu = containerGui.getMenu();
         ItemStack held = menu.getCarried();
         if (!held.isEmpty()) {
-            Slot under = GuiHelper.getSlotUnderMouse(containerGui).orElse(null);
+            Slot under = containerGui.hoveredSlot;
 
             for (Slot s : menu.slots) {
                 ItemStack stack = s.getItem();
-                if (!s.mayPickup(mc.player) || stack.isEmpty()) {
+                if (!s.isActive() || !s.mayPickup(mc.player) || stack.isEmpty()) {
                     continue;
                 }
 				Optional<StashResultAndTooltip> stashResultAndTooltip = getStashResultAndTooltip(stack, held);
@@ -103,7 +101,9 @@ public class ClientEventHandler implements ClientModInitializer {
         int y = containerGui.getGuiTop() + s.y;
 
         poseStack.pushPose();
-		poseStack.translate(0, 0, 300);
+		// Because of trinkets we need to increase this from the original 300
+		// Trinkets uses 310, so 330 was chosen as 320 was not enough as it cut the plus sign in half
+		poseStack.translate(0, 0, 330);
 
 		int color = stashResult == IStashStorageItem.StashResult.MATCH_AND_SPACE ? ChatFormatting.GREEN.getColor() : 0xFFFF00;
         if (stack.getItem() instanceof IStashStorageItem) {
@@ -116,7 +116,7 @@ public class ClientEventHandler implements ClientModInitializer {
 
     private static void renderSpecialTooltip(Minecraft mc, AbstractContainerScreen<?> containerGui, PoseStack poseStack, int mouseX, int mouseY, StashResultAndTooltip stashResultAndTooltip) {
         poseStack.pushPose();
-        poseStack.translate(0, 0, containerGui instanceof StorageScreenBase ? -100 : 100);
+        poseStack.translate(0, 0, 100);
         containerGui.renderTooltip(poseStack, Collections.singletonList(Component.translatable(TranslationHelper.INSTANCE.translItemTooltip("storage") + ".right_click_to_add_to_storage")), stashResultAndTooltip.tooltip(), mouseX, mouseY);
         poseStack.popPose();
     }
@@ -132,13 +132,13 @@ public class ClientEventHandler implements ClientModInitializer {
 		return Optional.empty();
 	}
 
-    private static Optional<StashResultAndTooltip> getStashResultAndTooltip(ItemStack potentialStashStorage, ItemStack potentiallyStashable, IStashStorageItem stashStorageItem) {
+	private static Optional<StashResultAndTooltip> getStashResultAndTooltip(ItemStack potentialStashStorage, ItemStack potentiallyStashable, IStashStorageItem stashStorageItem) {
 		IStashStorageItem.StashResult stashResult = stashStorageItem.getItemStashable(potentialStashStorage, potentiallyStashable);
 		if (stashResult == IStashStorageItem.StashResult.NO_SPACE) {
 			return Optional.empty();
 		}
 		return Optional.of(new StashResultAndTooltip(stashResult, stashStorageItem.getInventoryTooltip(potentialStashStorage)));
-    }
+	}
 
 	private record StashResultAndTooltip(IStashStorageItem.StashResult stashResult, Optional<TooltipComponent> tooltip) {}
 }