+++ src/main/java/net/p3pp3rf1y/sophisticatedcore/compat/CompatModIds.java
@@ -11,4 +11,5 @@ public class CompatModIds {
 	public static final String QUARK = "quark";*/
 	public static final String CHIPPED = "chipped";
 	public static final String LITEMATICA = "litematica";
+	public static final String AUDIOPLAYER = "audioplayer";
 }